Move schema/utils.ts into internals/graphql
