+++ __tests__/schema/uuid/abstract-exercise.ts
@@ -34,7 +34,7 @@ import {
   createTestClient,
   createUuidHandler,
 } from '../../__utils__'
-import { Model } from '~/schema/utils'
+import { Model } from '~/internals/graphql'
 import { EntityType } from '~/schema/uuid/abstract-entity/types'
 
 let client: Client
