+++ src/internals/graphql/index.ts
@@ -23,3 +23,4 @@ export * from './context'
 export * from './schema'
 export * from './requests-only-fields'
 export * from './resolver'
+export * from './utils'
