+++ src/schema/license/resolvers.ts
@@ -19,7 +19,7 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-import { Querys, TypeResolvers } from '~/schema/utils'
+import { Querys, TypeResolvers } from '~/internals/graphql'
 import { License } from '~/types'
 
 export const resolvers: TypeResolvers<License> & Querys<'license'> = {
