+++ src/schema/uuid/abstract-entity/resolvers.ts
@@ -20,7 +20,7 @@
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
 
-import { InterfaceResolvers } from '~/schema/utils'
+import { InterfaceResolvers } from '~/internals/graphql'
 
 export const resolvers: InterfaceResolvers<'AbstractEntity'> &
   InterfaceResolvers<'AbstractEntityRevision'> = {
