+++ src/schema/uuid/abstract-entity/types.ts
@@ -19,7 +19,7 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-import { Model } from '~/schema/utils'
+import { Model } from '~/internals/graphql'
 
 export enum EntityType {
   Applet = 'Applet',
