+++ src/schema/uuid/abstract-exercise/resolvers.ts
@@ -19,7 +19,7 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
-import { InterfaceResolvers } from '~/schema/utils'
+import { InterfaceResolvers } from '~/internals/graphql'
 
 export const resolvers: InterfaceResolvers<'AbstractExercise'> &
   InterfaceResolvers<'AbstractExerciseRevision'> = {
