+++ src/schema/uuid/abstract-exercise/utils.ts
@@ -19,8 +19,8 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
+import { PickResolvers } from '~/internals/graphql'
 import { SolutionDecoder } from '~/model'
-import { PickResolvers } from '~/schema/utils'
 
 export function createExerciseResolvers(): PickResolvers<
   'AbstractExercise',
