+++ src/schema/uuid/abstract-taxonomy-term-child/utils.ts
@@ -19,9 +19,9 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
+import { PickResolvers } from '~/internals/graphql'
 import { TaxonomyTermDecoder } from '~/model'
 import { resolveConnection } from '~/schema/connection/utils'
-import { PickResolvers } from '~/schema/utils'
 import { isDefined } from '~/utils'
 
 export function createTaxonomyTermChildResolvers(): PickResolvers<
