+++ src/schema/uuid/exercise-group/resolvers.ts
@@ -19,12 +19,12 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
+import { TypeResolvers } from '~/internals/graphql'
 import {
   ExerciseGroupDecoder,
   ExerciseGroupRevisionDecoder,
   GroupedExerciseDecoder,
 } from '~/model'
-import { TypeResolvers } from '~/schema/utils'
 import {
   createRepositoryResolvers,
   createRevisionResolvers,
