+++ src/schema/uuid/grouped-exercise/resolvers.ts
@@ -19,12 +19,12 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
+import { TypeResolvers } from '~/internals/graphql'
 import {
   ExerciseGroupDecoder,
   GroupedExerciseDecoder,
   GroupedExerciseRevisionDecoder,
 } from '~/model'
-import { TypeResolvers } from '~/schema/utils'
 import { createExerciseResolvers } from '~/schema/uuid/abstract-exercise/utils'
 import {
   createRepositoryResolvers,
