+++ src/schema/uuid/page/resolvers.ts
@@ -19,8 +19,8 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
+import { TypeResolvers } from '~/internals/graphql'
 import { PageDecoder, PageRevisionDecoder } from '~/model'
-import { TypeResolvers } from '~/schema/utils'
 import {
   createRepositoryResolvers,
   createRevisionResolvers,
